requirements.txt
et-xmlfile==1.1.0
openpyxl==3.0.9
pyserial==3.5
XlsxWriter==3.0.2
